refactor: 컴포넌트들을 components 패키지로 이동
