+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css";
-import TodoApp from "./TodoApp";
+import TodoApp from "./components/TodoApp";
 
 function App() {
   return <TodoApp />;
